+++ Example/Snabble/AppDelegate.swift
@@ -18,7 +18,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var dashboardViewController: DynamicViewController?
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-
         Asset.provider = self
 
         setupAppearance()
@@ -71,9 +70,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         let scannerViewController = AppScannerViewController(shop: shops.first!)
         let scannerNavigationViewController = UINavigationController(rootViewController: scannerViewController)
+        
+        let receiptsViewController = ReceiptsViewController()
+        let receiptsNavigationController = UINavigationController(rootViewController: receiptsViewController)
 
         let tabBarController = UITabBarController()
-        tabBarController.viewControllers = [dashboardViewController, scannerNavigationViewController, shopsViewController, accountNavigationController]
+        tabBarController.viewControllers = [dashboardViewController, shopsViewController, scannerNavigationViewController, receiptsNavigationController, accountNavigationController]
 
         window?.rootViewController = tabBarController