+++ Sources/UI/Receipts/ReceiptsListScreen.swift
@@ -1,0 +1,124 @@
+//
+//  ReceiptsListScreen.swift
+//
+//
+//  Created by Uwe Tilemann on 20.10.23.
+//
+
+import SwiftUI
+import SnabbleCore
+import Combine
+
+public extension PurchaseProviding {
+    var dateString: String? {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateStyle = .medium
+        dateFormatter.timeStyle = .short
+        dateFormatter.doesRelativeDateFormatting = true
+        return dateFormatter.string(for: date)
+    }
+}
+
+private struct BadgeCount: ViewModifier {
+    let badgeCount: Int
+    
+    func body(content: Content) -> some View {
+        if #available(iOS 15.0, *) {
+            content
+                .badge(badgeCount)
+        } else {
+            content
+        }
+    }
+}
+extension View {
+    func badgeCount(_ count: Int) -> some View {
+        modifier(BadgeCount(badgeCount: count))
+    }
+}
+
+public struct ReceiptsItemView: View {
+    public let provider: PurchaseProviding
+    public let image: SwiftUI.Image
+    
+    public init(provider: PurchaseProviding, image: SwiftUI.Image? = nil) {
+        self.provider = provider
+        self.image = image ?? Image(systemName: "scroll")
+    }
+    public var body: some View {
+        HStack {
+            self.image
+                .foregroundColor(.accentColor)
+                .padding(.trailing, 6)
+            VStack(alignment: .leading) {
+                Text(provider.name)
+                    .font(.headline)
+                Text(provider.dateString ?? "")
+                        .font(.footnote)
+            }
+            Spacer()
+            Text(provider.amount)
+                .font(.footnote)
+            Image(systemName: "chevron.right")
+                .font(.footnote)
+                .foregroundColor(.secondary)
+        }
+    }
+}
+
+private struct RefreshAction: ViewModifier {
+    let action: () -> Void
+    
+    func body(content: Content) -> some View {
+        if #available(iOS 15, *) {
+            content
+                .refreshable {
+                    action()
+                }
+        } else {
+            // TODO: we need something here
+            content
+        }
+    }
+}
+
+extension View {
+    func refreshAction(completion: @escaping () -> Void) -> some View {
+        modifier(RefreshAction(action: completion))
+    }
+}
+
+public struct ReceiptsListScreen: View {
+    @ObservedObject var viewModel: PurchasesViewModel
+    
+    public init() {
+        self.viewModel = PurchasesViewModel()
+    }
+
+    public var body: some View {
+        AsyncContentView(source: viewModel) { output in
+            VStack {
+                List {
+                    ForEach(output, id: \.id) { provider in
+                        ReceiptsItemView(provider: provider, image: viewModel.imageFor(projectId: provider.projectId))
+                        // provide a modifier for unloaded receipts here like:
+                        // .foregroundColor(provider.unloaded ? .primary : .secondary)
+                            .onTapGesture {
+                                viewModel.actionPublisher.send(provider)
+                            }
+                    }
+                    .listRowInsets(EdgeInsets(top: 10, leading: 16, bottom: 10, trailing: 16))
+                }
+                .listStyle(.plain)
+                .refreshAction {
+                    viewModel.load()
+                }
+            }
+        }
+        .onAppear {
+            viewModel.reset()
+        }
+        .badgeCount(viewModel.numberOfUnloaded)
+        .navigationTitle(Asset.localizedString(forKey: "Snabble.Receipts.title"))
+    }
+}